Remove a few more old objects, move Candidate object
