+++ app/controllers/ApplicationController.scala
@@ -19,9 +19,8 @@ package controllers
 import java.nio.file.Files
 
 import akka.stream.scaladsl.Source
-import model.Commands._
 import model.Exceptions.{ ApplicationNotFound, CannotUpdatePreview, NotFoundException, PassMarkEvaluationNotFound }
-import model.{ CreateApplicationRequest, ProgressStatuses }
+import model.{ CreateApplicationRequest, OverrideSubmissionDeadlineRequest, PreviewRequest, ProgressStatuses }
 import model.command.WithdrawApplication
 import play.api.libs.json.Json
 import play.api.libs.streams.Streams
@@ -49,8 +48,6 @@ object ApplicationController extends ApplicationController {
 }
 
 trait ApplicationController extends BaseController {
-  import Implicits._
-
   val appRepository: GeneralApplicationRepository
   val auditService: AuditService
   val applicationService: ApplicationService
