+++ app/repositories/sift/ApplicationSiftRepository.scala
@@ -18,11 +18,10 @@ package repositories.sift
 
 import factories.DateTimeFactory
 import model.ApplicationStatus.ApplicationStatus
-import model.Commands.Candidate
+import model._
 import model.EvaluationResults.Green
 import model.Exceptions.ApplicationNotFound
 import model.command.ApplicationForSift
-import model._
 import model.persisted.{ PassmarkEvaluation, SchemeEvaluationResult }
 import reactivemongo.api.DB
 import reactivemongo.bson.{ BSONArray, BSONDocument, BSONObjectID }
